backend/requirements.txt
Django==3.2.4
python-dotenv
biopython
Wikidata
psycopg2
nmslib==2.1.1
numpy
djangorestframework
djangorestframework-jwt
django-cors-headers
xmltodict
requests
django-background-tasks==1.2.5
pyldavis
gensim
spacy>=3.0.6,<3.1.0
wordcloud
pandas==1.2.4
SQLAlchemy==1.4.17
nltk==3.6.2
scispacy